requirements.txt
discord
requests
pyyaml